Disable mask computation for types containing enums
This means that if we don't have compiler information for a type that
has an enum in it somewhere, we'll refuse to project the enum value
because we don't know that we can do it correctly.

Also, disable the comparison of compiler-provided vs. locally-computed
masks for this case until this is fixed.