[SSADestroyHoisting] Don't fold into unrelated scopes.
If a load [copy] appears near the end of the scope protecting access to
another address and a destroy_addr of the loaded address appears
afterwards, don't fold the destroy into the scope.  The reason is that
doing so could allow a deinit which previously executed outside the
exclusivity scope to subsequently execute within it.