Synthesize read accessors to represent C++ [] operators
Previously, C++ [] operators were imported into Swift
by synthesizing a subscript containing a getter and a
setter. Particularly when the C++ operator returns a type
by reference, this creates unnecessary copies. A read
accessor can avoid this copy by yielding a reference to
the value. This change synthesizes are read accessor
instead of a getter when the return type is passed by
reference.