IRGen: address TODO for Windows static linking
This improves the static linking story on Windows by wiring up
information about the static vs dynamic build of the module into the
code generation for references to module external symbols.  We now can
indicate a module external symbol to be known local if it is expected to
be internalized due to static linking.  Even though the symbol may not
be defined in the same module, the definition will be provided locally
at link time.  This avoids needing to resolve a dllimport'ed symbol to a
local symbol at link time.