[move-function-value] Make sure that we propagate the debug scope from our original debug value into new undef debug values.
Otherwise if the move is in a different scope, the backend will think that we
have two different variables and emit the corresponding dwarf.

After I have this commit in and my alloc stack hoising commit (via a different
PR), I am going to be able to commit a full end <-> end test in lldb that values
are printed out appropriately as their lifetimes begin/end.