Disable a debugging assertion in the remote mirror MPE evaluation code
There are two sources of truth for the MPE spare bit mask in this code:

* The MPE reflection record which has a bit mask computed by the compiler.
  This should always be correct.

* A local computation within the RemoteMirror library.
  This is known to be incomplete, but is _supposed_ to fail cleanly
  when it hits a case it can't handle.

We need both of the above:  The first should always be accurate
so we obviously prefer it.  But old binaries don't have that data,
so we need the fallback computation.

When both of the above are available, they should agree, but apparently
the local computation still doesn't understand what it doesn't understand.

Resolves rdar://89360996