broaden the kinds of deinit contexts where we only warn in Swift 5
Unfortunately, for autoclosures (like 2nd arg of ||) we were saying that the
reference is still a hard error in a deinit. I didn't account for that when
initially patching the bug. For now, just go with the old context check
and put it before the checkedByFlow checking until I can improve this.