IRGen: loosen linkage on vtable stubs for static linking
The linkage inferred by internal linking is overly restrictive.  It uses
the "internal linkage type" which is for translation unit local symbols,
that is `static` storage.  These symbols participate in linking and need
to be given external storage but hidden visibility and no DLL storage in
static linking.  This permits them to participate in symbolic resolution
during linking but are made module local, not LLVM module local.