SIL: don't add effects attributes for storage decls.
Only for accessor functions.
This is important because the clang importer sets "readonly" attributes for a getter also on the storage decl, which would propagate to the setter.

This PR restores the old behavior which was changed by https://github.com/apple/swift/pull/40957.

Unfortunately I don't have a test case.

rdar://88876417