Update CMakeLists.txt
Explicitly invoke the script executor rather than the script as running a script is meaningless.  Shebangs are not a portable manner of indicating the script executor and worse yet do not guarantee portability on such systems either.  Ideally the shebang line would be removed from the script.

Thanks to @pcbeard for reporting the issue!