Cleaning up anchor finding in interpolated strings
The arglist didn't exist when I originally wrote this. Most of the
assertions were there to check my assumptions about the dyn-casts that
were being performed originally. In a non-asserts build, the dyn-casts
would not crash, but would emit the `await` insertion fix-it into the
wrong location. In an asserts build, it would crash instead.
I haven't heard about crashes, or misplaced fix-its from this for
roughly a year now, so I think it's safe to clean it up.

I'm keeping the `parent == nullptr` check since that would indicate an
issue in the implementation of `isEffectAnchor`.